data/requirements.txt
SQLAlchemy==2.0.25
PyMySQL==1.1.0
Faker==22.6.0
requests==2.31.0
python-dotenv==1.0.1
google-generativeai>=0.8.0
